refactor(winnig-lotto): winnig-lotto에서 prize 를 체크하도록 수정
