+++ src/main/java/lotto/domain/LottoTickets.java
@@ -4,9 +4,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import static java.util.stream.Collectors.collectingAndThen;
-import static java.util.stream.Collectors.toList;
-
 public class LottoTickets {
     private final List<LottoTicket> lottoTickets;
 
@@ -16,11 +13,5 @@ public class LottoTickets {
 
     public List<LottoTicket> lottoTickets() {
         return Collections.unmodifiableList(lottoTickets);
-    }
-
-    public LottoResult checkPrizesByWinningTickets(WinningLotto winningLotto) {
-        return lottoTickets.stream()
-                .map(winningLotto::matchPrize)
-                .collect(collectingAndThen(toList(), LottoResult::new));
     }
 }
